study2

+++ src/main/java/com/example/restfulwebservice/user/UserDaoService.java
@@ -20,9 +20,9 @@ public class UserDaoService {
     private static List<User> users = new ArrayList<>();
     private static int usersCount = 3;
     static {
-        users.add(new User(1, "kim", new Date(), "pass1", "701010-1111111"));
-        users.add(new User(2, "John", new Date(),"pass2", "701010-1111111"));
-        users.add(new User(3, "Mei", new Date(),"pass3", "701010-1111111"));
+        users.add(new User(1L, "kim", new Date(), "pass1", "701010-1111111"));
+        users.add(new User(2L, "John", new Date(),"pass2", "701010-1111111"));
+        users.add(new User(3L, "Mei", new Date(),"pass3", "701010-1111111"));
     }
 
     public List<User> findAll() {
@@ -31,7 +31,7 @@ public class UserDaoService {
 
     public User save(User user) {
         if (user.getId() == null) {
-            user.setId(++usersCount);
+            user.setId((long) ++usersCount);
         }
         users.add(user);
         return user;
